+++ packages/comp-ui-button/src/index.js
@@ -4,8 +4,10 @@
  * @Author: Guo Kainan
  * @Date: 2021-02-26 17:28:12
  * @LastEditors: Guo Kainan
- * @LastEditTime: 2021-02-26 17:28:33
+ * @LastEditTime: 2021-03-09 16:17:26
  */
 import Button from './Button.vue'
-
 export { Button }
+
+export * from './plugins/basic.js'
+export * from './plugins/animations.js'