+++ packages/vue-pixi-core/src/Renderer/directives/v-ticker.js
@@ -4,7 +4,7 @@
  * @Author: Guo Kainan
  * @Date: 2021-02-09 10:45:08
  * @LastEditors: Guo Kainan
- * @LastEditTime: 2021-03-08 19:42:40
+ * @LastEditTime: 2021-03-09 16:24:09
  */
 export default function vTicker (game) {
   let ticker = game.$app.ticker
@@ -20,7 +20,10 @@ export default function vTicker (game) {
       if (!handlerObj) {
         const handler = (time) => {
           if (typeof binding.value === 'function') {
-            binding.value(time, el, ticker, removeHandler)
+            binding.value(time, { 
+              el, ticker, 
+              stop: removeHandler 
+            })
           }
           else {
             console.error('The ticker handler must be a function!')