+++ packages/vue-pixi/src/index.js
@@ -4,13 +4,11 @@
  * @Author: Guo Kainan
  * @Date: 2021-02-26 17:38:46
  * @LastEditors: Guo Kainan
- * @LastEditTime: 2021-02-26 17:46:10
+ * @LastEditTime: 2021-03-09 16:19:17
  */
 export * from '@cmgl/vue-pixi-core'
 export * from '@cmgl/tween'
 
-export * from '@cmgl/active-container'
 export * from '@cmgl/button'
 export * from '@cmgl/progressbar'
-export * from '@cmgl/ellipse'
-export * from '@cmgl/rectangle'
+export * from '@cmgl/graphics'