+++ src/App.js
@@ -4,11 +4,11 @@
  * @Author: Guo Kainan
  * @Date: 2021-02-05 17:44:55
  * @LastEditors: Guo Kainan
- * @LastEditTime: 2021-03-08 17:06:26
+ * @LastEditTime: 2021-03-09 16:53:14
  */
 import { ref } from 'vue'
 
-import ActiveContainerView from '@/views/ActiveContainerView/ActiveContainerView.vue'
+import ButtonView from '@/views/ButtonView/ButtonView.vue'
 import ProgressBarView from '@/views/ProgressBarView/ProgressBarView.vue'
 import DragView from '@/views/DragView/DragView.vue'
 import ProjectionView from '@/views/ProjectionView/ProjectionView.vue'
@@ -16,7 +16,7 @@ import ProjectionView from '@/views/ProjectionView/ProjectionView.vue'
 export default {
   name: 'App',
   components: {
-    ActiveContainerView,
+    ButtonView,
     ProgressBarView,
     DragView,
     ProjectionView